renamed folders to supabase standards
