+++ my-app/middleware.ts
@@ -1,5 +1,5 @@
 import { type NextRequest } from 'next/server'
-import { updateSession } from './config/middleware'
+import { updateSession } from './utils/supabase/middleware'
 
 export async function middleware(request: NextRequest) {
   // update user's auth session
